pkg/processing: Let's fix the data race
There is a data race in ScyllaDB processor backend, as
it writes the ScyllaDB session variable, which is also read
by other goroutines.  Instead of lazily create a session,
the fix create a session during the creation phase.  Since
the gocql.Session is concurrency safe, this makes
the ScyllaDB concurrency safe without introducing the
sync.Mutex based serialization.